Move IGL_GdiPlus.RenderTargetWrapper to own file
